+++ CHANGELOG.md
@@ -7,11 +7,12 @@ This file is similar to the format suggested by [Keep a CHANGELOG](https://githu
 
 ## Unreleased
 - [Feature] Add Github action to automatically update CHANGELOG.md based on PR comment ([#1399](https://github.com/optimizely/oui/pull/1399))
+- [Release] Update Color design tokens and rename amber to yellow ([#1398](https://github.com/optimizely/oui/pull/1398))
 
 ## 47.0.0 - 2020-08-07
 - [Release] Remove `word-break` helper class in favor of `word-break--all` ([#1394](https://github.com/optimizely/oui/pull/1394))
 - [Feature] Add `hasClearButton` to **DatePicker** and **DateRangePicker** ([#1396](https://github.com/optimizely/oui/pull/1396))
-- [Feature] Add `leftIcon` and `rightIcon` to **Link** ([#1393](https://github.com/optimizely/oui/pull/1393)) 
+- [Feature] Add `leftIcon` and `rightIcon` to **Link** ([#1393](https://github.com/optimizely/oui/pull/1393))
 - [Patch] Replaced `name` and `secondaryText`'s string requirement to node for **Dropdown** ([#1395](https://github.com/optimizely/oui/pull/1395))
 
 ## 46.13.0 - 2020-08-05